230120 김형호
코드 업데이트

+++ AR41DX/AR41Engine/Include/Resource/Mesh/MeshManager.cpp
@@ -136,7 +136,7 @@ bool CMeshManager::Init()
 	CreateMesh(nullptr, MeshType::Static, "SpherePos",
 		&vecSpherePos[0], sizeof(Vector3), (int)SphereSize,
 		D3D11_USAGE_DEFAULT, D3D11_PRIMITIVE_TOPOLOGY_TRIANGLELIST,
-		&vecSphereIndex[0], sizeof(int), vecSphereIndex.size(),
+		&vecSphereIndex[0], sizeof(int), (int)vecSphereIndex.size(),
 		D3D11_USAGE_DEFAULT, DXGI_FORMAT_R32_UINT);
 
 	return true;
